+++ lua/obsidian/actions.lua
@@ -68,7 +68,7 @@ M.smart_action = function()
     return legacy and "<cmd>ObsidianTags<cr>" or "<cmd>Obsidian tags<cr>"
   elseif has_markdown_folding() and api.cursor_heading() then
     return "za"
-  elseif api.cursor_checkbox() or Obsidian.opts.checkbox.create_new then
+  elseif Obsidian.opts.checkbox.enabled and (api.cursor_checkbox() or Obsidian.opts.checkbox.create_new) then
     return legacy and "<cmd>ObsidianToggleCheckbox<cr>" or "<cmd>Obsidian toggle_checkbox<cr>"
   else
     return "<CR>"