+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed markdown backlinks, so they are properly located with regex and respect to anchoring
+- Template substitutions revert to support string values.
 - Cmdline completion will work for multiple arguments.
 - `:Obsidian open`'s argument completion will work.